pyproject.toml
[project]
name = "pjecz-casiopea-flask"
version = "0.1.0"
description = "Plataforma de administración del sistema de citas."
authors = [
    {name = "Guillermo Valdes",email = "guillermo@movimientolibre.com"}
]
license = {text = "AGPL3"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-moment (>=1.0.6,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-socketio (>=5.5.1,<6.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "google-cloud-secret-manager (>=2.23.3,<3.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "hashids (>=1.3.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "rq (>=2.3.2,<3.0.0)",
    "sendgrid (>=6.11.0,<7.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "wtforms (>=3.2.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
pylint = "^3.3.6"
pylint-sqlalchemy = "^0.3.0"
pytest = "^8.3.5"
requests = "^2.32.3"
faker = "^37.1.0"

[tool.poetry.scripts]
cli = "cli.app:cli"

[tool.black]
line-length = 128

[tool.isort]
line_length = 128
profile = "black"
